Реорганизация explode.*
